requirements.txt
asgiref==3.3.4
autopep8==1.5.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-ckeditor==6.1.0
django-debug-toolbar==3.2.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
Unidecode==1.2.0
xlrd==2.0.1
xlwt==1.3.0
